Ignore test results for gnmi_server
The Debian files attempt to run make check for the gNMI server, however,
it fails on a build machine, because it expects the Redis server to be
available and listening on localhost port 6379.

For now, the result of the test is ignored, and we have to find a way to
fix this so that the tests work even on a build machine.